requirements.txt
Django==1.8
pkg-resources==0.0.0
django-autofixture==0.9.2
django-encrypted-fields==1.1.2
loremipsum==1.0.5